requirements.txt
openai>=1.0.0
openai-agents>=0.0.17
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
gradio>=4.0.0